% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_fit.R
\name{count_fit}
\alias{count_fit}
\title{Fit different distributions to count data}
\usage{
count_fit(wf)
}
\arguments{
\item{wf}{A whitefly count dataframe.}
}
\value{
List of distribution parameters for diferent fits.
        and histogram data for plotting.
}
\description{
Fit different distributions to count data
}
\examples{
# fit.list <- count.fit(plant.count)
}
