% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_aggregate.R
\name{efficacy_aggregate}
\alias{efficacy_aggregate}
\title{Efficacy calculations as ratio of count means, geometric means, or medians}
\usage{
efficacy_aggregate(wf, control = "COL1468")
}
\arguments{
\item{wf}{A whitefly count dataframe.}

\item{control}{Control genotype string.
Genotype Denominator in efficacy: e = u1/u0.}
}
\value{
A dataframe with efficacy calculated per clone
}
\description{
Efficacy calculations as ratio of count means, geometric means, or medians
}
\examples{
# efficacy.df <- efficacy.aggregate(wf.per.plant ~ control = "COL1468")
}
