% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficacy_list.R
\name{efficacy_list}
\alias{efficacy_list}
\title{Vertical to horizontal data reformatting for efficacy calculations
just a wrapper for \code{reshape2::dcast} from \code{reshape2}}
\usage{
efficacy_list(ef)
}
\arguments{
\item{ef}{efficacy dataframe from \code{efficacy.aggregate()}}
}
\value{
List of wide formated dataframes of efficacy by mean,
        geometric mean median
}
\description{
Vertical to horizontal data reformatting for efficacy calculations
just a wrapper for \code{reshape2::dcast} from \code{reshape2}
}
\examples{
# ef.by.measure <- efficacy.list(efficacy.df)
}
