% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nb_glm.R
\name{fit_nb_glm}
\alias{fit_nb_glm}
\title{Fit MASS Negative binomial General Linear Model for whitefly counts}
\usage{
fit_nb_glm(formula, wf)
}
\arguments{
\item{formula}{formula object specifiying the GLM}

\item{wf}{A whitefly count dataframe.}
}
\value{
List with model fit, lsmeans and posthoc groups
}
\description{
Fit MASS Negative binomial General Linear Model for whitefly counts
}
\examples{
# fit.nb.glm(nymphs ~ year + leaf, wf.per.leaf)
}
