% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgamma.R
\name{fitgamma}
\alias{fitgamma}
\title{Equivalent to `MASS::fitdistr(x, densfun = "gamma")`, where x are first
rescaled to the appropriate scale for a gamma distribution.
Useful for fitting the gamma distribution to 
data which, when multiplied by a constant, follows this distribution}
\usage{
fitgamma(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
MASS::fitdistr with scaling multiplier
}
\description{
Equivalent to `MASS::fitdistr(x, densfun = "gamma")`, where x are first
rescaled to the appropriate scale for a gamma distribution.
Useful for fitting the gamma distribution to 
data which, when multiplied by a constant, follows this distribution
}
\examples{
fitgamma(1:1000)
}
