% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Geometric mean of a vector ignoring NAs.}
\usage{
geometric_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector.}
}
\value{
Geometric mean of  \code{x} ignoring NAs.
}
\description{
Geometric mean of a vector ignoring NAs.
}
\examples{
geometric.mean(1:1000)
}
