% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Emulate ggplot color palette.}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors to generate}
}
\value{
vector of hexadecimal color strings

# @examples
gg_color_hue(length(levels(wf$clone)))
}
\description{
Emulate ggplot color palette.
}
