% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_count.R
\name{leaf_count}
\alias{leaf_count}
\title{Counts per leaf}
\usage{
leaf_count(wf)
}
\arguments{
\item{wf}{A dataframe of whitefly count counts per picture.}
}
\value{
Dataframe containing whitefly counts per leaf
}
\description{
Counts per leaf
}
\examples{
# per.leaf <- leaf.count(my.wf)
}
