% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_count.R
\name{plant_count}
\alias{plant_count}
\title{Counts per plant}
\usage{
plant_count(wf)
}
\arguments{
\item{wf}{A dataframe of whitefly counts per picture}
}
\value{
Dataframe containing whitefly counts per plant
}
\description{
Counts per plant
}
\examples{
# per_plant <- plant.count(my_wf)
}
