% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_check_cor.R
\name{plot_check_cor}
\alias{plot_check_cor}
\title{Plot wide formated dataframe correlations per clone
with point labels, asumes plotted points to be ~ 10
more points would be extremely unreadable}
\usage{
plot_check_cor(check_wide, main = "Correlation", ...)
}
\arguments{
\item{check_wide}{wide formatted data by clone for checks only}

\item{main}{plot title}
}
\value{
\code{pairs} plot of correlation
}
\description{
Plot wide formated dataframe correlations per clone
with point labels, asumes plotted points to be ~ 10
more points would be extremely unreadable
}
\examples{
# plot_check.cor(eplot_check.cor)
}
