% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count_fit.R
\name{plot_count_fit}
\alias{plot_count_fit}
\title{Plot Probability Density over histogram to appreciate fit}
\usage{
plot_count_fit(x, ...)
}
\arguments{
\item{x}{A \code{count.fit()} list}
}
\value{
plot_
}
\description{
Plot Probability Density over histogram to appreciate fit
}
\examples{
# plot_count.fit(count.fit(wf.per.leaf))
}
