% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cross_anova .R
\name{plot_cross_anova}
\alias{plot_cross_anova}
\title{Plot ANOVA per cross}
\usage{
plot_cross_anova(posthoc, exp_allowed)
}
\arguments{
\item{exp_allowed}{list of vectors containing indices of experiments
to merge in order to make ANOVA analysis}

\item{postcoc}{\code{multcomp} common letter display \code{cld} object from
posthoc comparison in \code{plot_fit.nb.glm()}}
}
\value{
Nothing
}
\description{
Plot ANOVA per cross
}
\examples{
# plot_cross.anova(posthoc, exp_allowed)
}
