% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cross_nb_fit.R
\name{plot_cross_nb_fit}
\alias{plot_cross_nb_fit}
\title{Plot nb GLM per cross}
\usage{
plot_cross_nb_fit(posthoc, exp_allowed, breaks = NULL)
}
\arguments{
\item{posthoc}{\code{multcomp} common letter display \code{cld} object from
posthoc comparison in \code{plot_fit.nb.glm()}}

\item{exp_allowed}{list of vectors containing indices of experiments
to merge in order to make nb GLM analysis}
}
\value{
Nothing
}
\description{
Plot nb GLM per cross
}
\examples{
# nb.fit(posthoc, exp_allowed)
}
