% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ef_cor.R
\name{plot_ef_cor}
\alias{plot_ef_cor}
\title{Plot wide formated dataframe correlations per clone
in arc sine scale}
\usage{
plot_ef_cor(ef_par, main = "Correlation", ...)
}
\arguments{
\item{main}{plot title}

\item{ef.par}{wide formatted data by clone}
}
\value{
\code{pairs} plot of correlation
}
\description{
Plot wide formated dataframe correlations per clone
in arc sine scale
}
\examples{
# plot_wide.cor(efficacy_list(ef)$mean)
}
