% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ef_density.R
\name{plot_ef_density}
\alias{plot_ef_density}
\title{Plot of efficacy probability density
just a wrapper for \code{reshape2::dcast} from \code{reshape2}}
\usage{
plot_ef_density(ef, type, group)
}
\arguments{
\item{ef}{efficacy dataframe from \code{efficacy.aggregate()}}
}
\value{
List of wide formated dataframes of  efficacy by mean,
        geometric mean, median by clone for each experiment
}
\description{
Plot of efficacy probability density
just a wrapper for \code{reshape2::dcast} from \code{reshape2}
}
\examples{
# ef_by_measure <- plot_ef_density(efficacy.df)
}
