% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit_nb_glm.R
\name{plot_fit_nb_glm}
\alias{plot_fit_nb_glm}
\title{Plot negative binomial model fit}
\usage{
plot_fit_nb_glm(wf, nb_fitted, type = "violin", xmax = 15000, cld = TRUE)
}
\arguments{
\item{wf}{A whitefly count dataframe.}

\item{type}{\code{"CI"} Confidence interval dotplot
\code{"density"} probability density joyplot}

\item{np.fitted}{a \code{fit.nb.glm()} list.}
}
\value{
Nothing
}
\description{
Plot negative binomial model fit
}
\examples{
# plot_fit_nb_glm(wf_per_leaf,
#                 fit_nb_glm(nymphs ~ year + leaf, wf.per.leaf),
#                 type = "density")
}
