% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gg_infestationXclone.R
\name{plot_gg_infestationXclone}
\alias{plot_gg_infestationXclone}
\title{Plot infestation X clone from GLM}
\usage{
plot_gg_infestationXclone(cld)
}
\arguments{
\item{cld}{\code{multcomp} common letter display \code{cld} object from
posthoc comparison in \code{plot_fit.nb.glm()}}
}
\value{
Nothing
}
\description{
Plot infestation X clone from GLM
}
\examples{
# plot_gg.infestationXclone(nb.glm.fit$posthoc)
}
