% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infestationXclone.R
\name{plot_infestationXclone}
\alias{plot_infestationXclone}
\title{Plot infestation X clone interaction as boxplot}
\usage{
plot_infestationXclone(wf)
}
\arguments{
\item{wf}{A whitefly count dataframe.}
}
\value{
Nothing
}
\description{
Plot infestation X clone interaction as boxplot
}
\examples{
# plot_infestationXclone(wf_per_leaf)
}
