% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wide_cor.R
\name{plot_wide_cor}
\alias{plot_wide_cor}
\title{Plot wide formated dataframe correlations per clone
in log scale}
\usage{
plot_wide_cor(wf_wide, main = "Correlation", ...)
}
\arguments{
\item{wf_wide}{wide formatted data by clone}
}
\value{
\code{pairs} plot of correlation
}
\description{
Plot wide formated dataframe correlations per clone
in log scale
}
\examples{
# plot_wide.cor(efficacy.list(ef)$mean)
}
