% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wf.R
\name{read_wf}
\alias{read_wf}
\title{Read whitefly count table}
\usage{
read_wf(file, ...)
}
\arguments{
\item{file}{A filename for the whitefly counts per picture.}
}
\value{
Dataframe containing whitefly counts
}
\description{
Read whitefly count table
}
\details{
Whitefly counts per picture. Each picture must have a unique combination of
experiment + replicate + plant + pic
experiment description consists of:
exp.year + exp.cross + exp.propagation + exp.substrate
}
\examples{
# my.wf <- read.wf("whitefly_counts.tab")
}
