% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_levels.R
\name{remove_levels}
\alias{remove_levels}
\title{Remove levels form factors of subsetted dataframes}
\usage{
remove_levels(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
Dataframe with excess levels removed from factors
}
\description{
Remove levels form factors of subsetted dataframes
}
\examples{
# wf2000 <- remove.levels(plant.count[plant.count$year = 2000, ]
}
