% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_aggregate.R
\name{wf_aggregate}
\alias{wf_aggregate}
\title{Aggregate whitefly data by single criterion}
\usage{
wf_aggregate(wf, x = "clone", y = "nymphs", by_stat = "mean")
}
\arguments{
\item{wf}{A whitefly count dataframe.}

\item{x}{criterion column for aggregation.}

\item{y}{variable column for aggregation.}

\item{by_stat}{column for ordering levels for graphs.}
}
\value{
Dataframe of aggregates by specified {col} 
        values and levels ordered by \code{by_stat}
}
\description{
Aggregate whitefly data by single criterion
}
\examples{
# wf_aggregate(my.wf,x="clone", y="nymphs", by_stat ="geometric.mean" )
}
