% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_wide.R
\name{wf_wide}
\alias{wf_wide}
\title{Vertical to horizontal data reformatting and aggregation for whitefly counts
just a wrapper for \code{reshape2::dcast} from \code{reshape2}}
\usage{
wf_wide(formula, wf, fun = geometric_mean)
}
\arguments{
\item{formula}{formula object specifiying the \code{reshape2::dcast} reformatting}

\item{wf}{A whitefly count dataframe.}

\item{fun}{function for aggregation deafult \code{geometric.mean}}
}
\value{
A wide formated dataframe
}
\description{
Vertical to horizontal data reformatting and aggregation for whitefly counts
just a wrapper for \code{reshape2::dcast} from \code{reshape2}
}
\examples{
# clone.means <- wf.wide(nymphs ~ clone,fun = mean)
}
